REFACTOR: auth base test case
