+++ tests/whist/server/api/game/test_create.py
@@ -1,22 +1,8 @@
-import unittest
-
-from starlette.testclient import TestClient
-
-from whist.server import app
+from tests.whist.server.api.game.base_token_case import TestCaseWithToken
 from whist.server.database import db
 
 
-class CreateGameTestCase(unittest.TestCase):
-    def setUp(self) -> None:
-        self.client = TestClient(app)
-        self.login_creds = {'username': 'marcel', 'password': 'abc'}
-        _ = self.client.post(url='/user/create', json=self.login_creds)
-        token = self.client.post(url='/user/auth/', data=self.login_creds).json()['token']
-        self.headers = {'Authorization': f'Bearer {token}'}
-
-    def tearDown(self) -> None:
-        db.game.drop()
-        db.user.drop()
+class CreateGameTestCase(TestCaseWithToken):
 
     def test_post_game(self):
         data = {'game_name': 'test', 'password': 'abc'}
